vdl: fix getReflectTypeForVdl function
We want to make sure that if we generate java reflect type
for VDL type then we can reconstruct VDL type from the
generated java reflect type.

If kind of VDL type is array, enum, struct or union we
just can't build the java reflect type so that we can
reconstruct the VDL type back. That is why we need to
annotate generated java class for array, enum, struct, union.

The case of VDL wrappers just doesn't work for the same reason.
So we move the VDL wrappers case to a separate function and
build only top-level java type to be able to decode into vdl.Value.

Change-Id: Ifaf4a9d1406863f5226b99fd17c7cbf338d78b9e