Java: add the Invoker interface and cleanup rpc code
We currently create a new invoker for every method invocation, which
is costly.  The invoker interface allows us to share the same invoker
accross calls; also, it brings us on par with the Go implementation.

Also simplified multiple return values handling and renamed VDLInvoker
to ReflectInvoker.

Removes UniversalServiceMethods, as it's no longer used.

Added a bunch of tests to test the new behavior.

Change-Id: I1433e21c3e3be8f8a4f123965f802c8f08c2163b