Share Java tests with Android
The purpose of this CL is to share tests between "regular" Java and
Android environments. In order to implement this, it introduces a new
"tests" project that is a sibling of "lib" and "android-lib".

We place all tests into the "tests/src/main/java" (*not* src/test/java).
The test project has a compile dependence on lib. lib in turn has a
testCompile dependence on tests. The sole remaining test in 'lib' is a
test suite that runs all the test classes from 'tests'. Because of this,
you can run ./gradlew :lib:tests from the java project root as normal.

The android-lib project now has a testCompile dependence on tests. We
add a test suite called 'AllTests' to the android-lib. This AllTests
suite scans the Android APK and collects up all the tests to run.

There are miscellaneous changes to advance the Android gradle plugin
version and build tools version. The latter will require users to
install the newer version (in Android Studio, this means clicking a
button on the error that occurs if the newer version is not already
installed).

Change-Id: I2ecb5ce929bf39695673579c1eb3f1da5f613022