gradle-plugin: no longer require vanadium to be installed
This CL removes the requirement that users of the VDL plugin have V23
installed before they start. This works by shipping the vdl binary along
with the gradle plugin. Each time the vdl plugin is started, we extract
the vdl binary to a temporary location in the project's build directory.
The vdl binary itself is stored in the root of the  gradle plugin JAR
file.

Change-Id: I139810eb53e23a5f1f07d99c8c1f2903708f91f2