android: move blessing fetching from identityd into a library
Also, rewrite the fetching code to use the new http-based mechanism.

The ultimate goal is for app developers to be able to use this library
instead of relying on blessings from the account manager (which
requires a separate installation).

If the account manager is available, however, we will go through that
blessing path.

Change-Id: I67e34334fafd2f076b5537cc023ae5c509dac819