syncbase: Removing the constructor that took a string as it was confusing to whether it meant a single row or a prefix.
With this CL we now have three different named constructors
(static factory methods on RowRange class) that can be
used to create RowRanges given limits, prefixes or single rows.

Constructors for RowRange and PrefixRange were made to be
package-private so that there is only a single way to
created RowRanges through the factory methods.

Change-Id: Idd4b680e27bb5a78381c5a112811aeed48393722