syncslides: Add a FAB to create a new presentation.
Add a floating action button (FAB) to create a new presentation.
This change also replaces the main content layout with a
CoordinatorLayout.  CoordinatorLayout has good interactions between
the FAB and any snackbar that we will add later.

The icon for the FAB is just a placeholder until James produces
the real one.

There is an onClickListener for the FAB that just logs right now.
The intention is that Dan will hook this up to the Google Drive
import flow he has been working on.

Change-Id: I66a9132e070949df58cf7490005191be055d92aa