Java: Bluetooth protocol support for Android
Allows (Android) servers to listen on Bluetooth endpoints and
(Android) clients to connect to them.

The protocol name is "bt".

The (bluetooth) address format is:

      <MAC Address>/<Port Number>

e.g.,
      AA:BB:CC:DD:EE:FF/20

MAC Address is the address of the local bluetooth device and the
port number is one of the available 31 bluetooth ports.  As with
TCP, server can mount itself on:

     <MAC Address>/0

, in which case the first available bluetooth port will be used.

Change-Id: Ic49712ea8436e94a789f948bcb434afb9c327ccc