syncslides: Mock list of presentations.
Add a RecyclerView to show the list of presentations.  We had thought
about using a simple GridView since the number of presentations will
be small, but the docs suggested using RecyclerView if the list will
change dynamically (which it will).

The list of presentations is hardcoded for now.  There's more work to
do on tweaking the card layout, too.

Screenshot: https://screenshot.googleplex.com/VYSuk5mXWo2.png
Change-Id: Ib6aa10782484dfb82d359f7b797d823a2671dadd