syncslides: DB interface to make it easy to fake syncbase.
The DB interface hides the details of syncbase from the UI layer
to make it easy to implement a fake version.

We want only one instance of syncbase for the application, so
it lives in a singleton.  In the long run, it would be better for
syncbase to run as a service.  This change is compatible with that
plan.

The FakeDB implementation currently implements only the functionality
necessary for the DeckListActivity.  It has a background thread to
demonstrate how the data in the DB can change dynamically and have
that update the UI.

Change-Id: I28615df55c79f1bd64178f62d1c11b09ebd47636