syncslides: Layout for Audience/Presenter view.
Both the audience and presenter views are combined into
a single NavigateFragment.  The idea is that NavigateFragment
will take a parameter specifying whether it is an audience or
presenter.  That will make certain buttons (in)visible.

This change doesn't do any of that.  Everything that might change
between audience and presenter is invisible. It hardcodes the slide
content.  It doesn't have any click listeners.  This change is
mostly to get the majority of the layout in place so that other
stuff can be added incrementally.

Added a click listener to the SlideListAdapter so there is a way
to navigate to this new view.

Change-Id: I9d7f595d8aa3f481fb7f07d863cf1d55dc119f64