syncslides: Immersive mode for navigation view.
Immersive mode hides the status bar (wifi, battery),
toolbar (app-specific menu), and Android navigation softkeys (back, home,
recents).  When the user is viewing the NavigationFragment, we want
maximum screen real estate, so it is important to hide all of these
system UI elements.  We use the "sticky" version of immersive mode
so the user can access the status bar and navigation softkeys by
swiping from the top or bottom.

We want to enter immersive mode whenever the NavigationFragment starts
and exit immersive mode when it stops (e.g. user goes back to slide list).
However, this doesn't just work.  Going back to the slide list causes
the slide list's toolbar to be shifted down.  Between the toolbar and the
status bar, there is a gray strip that is the same height as the status
bar.  So this change works around this mis-render by having the content
of the slide list go under the toolbar.

- Cleanup our AndroidManifest.xml now that we can navigate through
  all our views.
- Add NavigateFragment to the back stack so the back button takes
  the user back to the slide list.

Change-Id: Icfbb13ad34f1a7a9552ede5730f0fa2c4f61d2e5