syncslides: Hook up navigation icons.
Touching on the prev/next icons or thumbnails will go to the
prev/next slide accordingly.  The NavigateFragment also starts
at the slide the user touched on the SlideListFragment.

The slidelist (layers) icon pops the backstack to get back to
the SlideListFragment.

The note content is populated from the DB now.

Added a Role enum so the NavigateFragment can show certain buttons
to the presenter and other buttons to the audience.

Added an asynchronous DB.getSlides() to make it easier to populate
the images if the data is not already loaded (e.g. fragment is restored).

Change-Id: I5db007932b2f026ccb67d6f6acb5ed3ca955b5d0