syncslides: Fix NPE.
Fix for https://github.com/vanadium/java/issues/5

onPause causes mAdapter to be unusable, so we need to create a new
one when the fragment resumes.  However, onPause can still show the
fragment in the background, so we might get weird behavior if
the adapter is stopped and started while the fragment is visible.
Therefore, this change uses onStart/onStop instead of onResume/onPause.

Change-Id: Idd28982c1363bf5a4fc173e731b2e06e1c38b282