syncslides: Better handling of EditText focus.
We need an explicit action for the user to say that they are done
editing the notes.  The notes can't be both multiline and have
the "done" button in the soft keyboard.  This change adds a
"Save" action to the ActionBar when the user is editing.  This
action hides the keyboard and resets the view to immersive.

There is still a big problem left to solve: touching one of the
navigation icons advances to the next/prev slide but does not
save nor hide the keyboard.

But this change is a big improvement over the previous behavior
where you couldn't ever stop editing, so I figured this was
a good checkpoint.

Change-Id: I8c1cd99dec680840ce36d86b3ce76d8f90489dbe