syncslides: Add audience sync FAB
If the audience member starts the presentation by clicking on a slide
card, or if the audience member switches slides or enters notes, the
member becomes unsynced and does not change slides when the presenter
does. This is visible via a sync icon FAB. When the audience member
clicks the FAB, it disappears and the member is synced with the
presenter again.

Screenshot: https://screenshot.googleplex.com/2wKo5NBR3Ag.png
Change-Id: Iba1be5dc8c99c57d3c2fc95abddfab11e1608eb2