syncslides: Add a missing DB.init().
We need to initialize DB at the beginning of each activity.
We didn't see this as a problem since the only activity that
currently starts PresentationActivity already calls DB.init().
If that were to change (e.g. PresentationActivity started
by a URL handler), syncbase would not be initialized when
the activity tried to access it.

Change-Id: Ib98ad8fb18607f3bf626b80b4040c3215d2fae33