Split vanadium maven artifact
Created a new maven artifact: vanadium-without-natives.
Projects like android-lib should depend on this.

With this patch I observed the build time for the following command:

syncslides$ ./gradlew --daemon assembleDebug

fall from 33s to 18s for an incremental build with a single-line change.
Also package size fell from 24mb to 7mb because we no longer include
both Linux and MacOSX natives in the APK!

vanadium-android now has a transitive dependence on
vanadium-without-natives, so it's no longer necessary to explicitly
declare a dependence on vanadium from Android projects.

Change-Id: I61192d0a12b9bc90fed7bd6cb65ea91801bfe695