syncslides: Refactor and rename Slide and Deck.
We had two classes for each of Slide and Deck, making it really annoying
to work with both types in the same file because you have to provide
the fully qualified type everywhere.  I renamed the VDL types to
VSlide and VDeck.

Also, I started using DeckImpl instead of the very similar one in
SyncbaseDB.java.  I similarly added SlideImpl.  This should make
SyncbaseDB.java not so monolithic.

Change-Id: I12215b567b18f8926fbfaee15e2f39d9369c0294