syncslides: Keep audience sync state
When an audience member switches back and forth between the navigate
fragment and fullscreen fragment, keep the sync state consistent.
Also, swap between the navigate fragment and fullscreen fragment
instead of adding the navigate fragment to the back stack and popping
the fullscreen fragment (and modify back stack so that it still works
normally on back presses.)

Move the sync state to the activity rather than pass it between
fragments.

To run this cl after patching it, change role on line 111 of
DeckListAdapter to Role.AUDIENCE and select a FakeDB deck.

Fixes #22, fixes #24

Change-Id: Ie688fb34df02c1a8d2cf15d4a95f2a49b817f4c3