syncslides: Fix deck object creation and use.
In the course of changing the proto served by PresentationActivity from
'Description' to 'VDeck', and testing different scenarios with
presentation begin/end, found bugs rooted in inconsistent construction
of DeckImpl (we aren't using @Nullable or @Optional, so are at more risk
of NPE).  So consolidated construction to one place, and added
consistent placeholder data to avoid nulls.

Also, fixed a problem in that rotating the presenting phone could result
in hangs on subsequent gets from clients (other phones) because the
underlying v23 code caches the endpoint - and the endpoint can change in
a rotation cycle, because the server lives only as long as the
PActivity.  Filed https://github.com/vanadium/issues/issues/812
to track what appears to be a problem with endpoint caching.

Also fixed a problem where we sometimes removed a discovery listener
while the off-UX discovery thread was still in the course of being
stopped.

Change-Id: Ibc90f4e08bda351788e52dd26419254293c7bfc2