syncslides: UI fixes
Take care of a few problems:
- Currently, the actionbar and soft keyboard stay visible if the user is
typing notes and then presses on a different UI element instead of
saving first. We add a custom onClickListener that saves the notes and
makes the action bar and keyboard disappear.
- Added the slide list to the backstack for presenter mode so that the
presenter can press the slide list button or back button properly.
- If the user presses the back button from the questions fragment,
remove the action bar and bottom bar.
- Put a background on the next / prev thumbs. Sometimes the slide list
is visible behind them.

Change-Id: I9610507840ff0cf419808cbe3f1b0121b1732d0b