syncslides: Sign-in to Google to get user's name
The mechanism is as follows:

   - we add a new starting activity that prompts the user to sign in
     using one of his google.com accounts; this account for now is
     the only account used by the app.
   - we get the user profile corresponding to the google.com account and
     extract first/last name from it (in the future, profile pic etc.).
   - this account/profile is stored in shared preferences for other
     activities to use (current users are NavigationalDrawerFragment
     and V23Manager, for minting the blessings).
   - the activity then invokes the DeckChooserActivity and things
     are the same as before.

I needed to split this into a separate activity so that this entire
sign-in path is guaranteed to happen before anything else.  It's a
one-time thing, after the first sign-in things are saved in the app.

Below is the screenshot with the filled-in user info on the
navigation drawer.

https://drive.google.com/file/d/0B6WJqpR8KjWsS0pRRUNSaTFwV1E/view?usp=sharing

Finally, note that this may have been easier if we used the Google+ Sign-In
button, but it requires pulling GooglePlayServices libraries and also
registering the app on the console with a fixed sign-in key, which is
a pain for everyone as it would no longer be possible to run in debug
mode etc.

Change-Id: Ib9bf42d017bd2c88b395e559b6cc0d5d5ce019a1