syncslides: Fix navigate onCreateView slide num
Right now the current slide number is 0 when we create the view, and
when we sync we set the user slide number to it, which means we
always initialize the fragment at slide 0. Fix this by setting the
current slide to the number that we put in the bundle, and syncing
the slide number to that.

Change-Id: I08296f3b98d8c38800bfbc09e6918f73cd7c9caa