syncslides: Read the user's name from Contacts.
We're going to read the user's name from Contacts rather
than trying to read it from the OAuth account.  OAuth requires
an internet connection to set up.  While that is a one-time
thing.  It might be necessary to reset the app's storage while
at Founder World.  In that case, we wouldn't be able to go
through the OAuth flow again (no internet).

Also, rename the app to "SyncSlides" since we're getting close
to having a final version.

Change-Id: I3ba941a928b050c43ea41eed0f0938ad89b75706