syncslides: attempt at optimizing the import path
In particular, don't encode/decode bitmaps when loading slides.

The full-res slides (300KB) take about 1.5 seconds each to get
written into the DB.  Most of this time is spent writing the data
into a blob store.  This should be separately investigated and
optimized.

For now, calling an async import method to keep UI responsive.
It means that the deck will become almost instantly visible, but
the slides in it will keep loading.  Our UI nicely handles
this scenario.

Change-Id: Icb6a31b653ec92447c205a36821027a1ab9768d6