only modify current project's VDL input path
Before this change, the plugin would modify the inputPaths field of the
VDL extension of all projects that it encountered in its recursive
descent. This meant that, in a multi-project build, a given project may
have its inputPaths field altered many times.

This change causes that logic to happen only once per project. This is
achieved by moving the extension manipulation into the prepareVdl task.

Change-Id: Ic8c3bd8746c7f7e32e84860dd50334db9e636401