fix over-eager evaluation of various VDL paths
In Gradle, the values of extension fields is not finalized until project
evaluation. This means that we must defer reading the values as long as
possible. To achieve this, we specify paths either in closures (e.g.
delete { project.vdl.outputPath }, this closure gets evaluated only when
the task is going to be performed), or in project.afterEvaluate.

Change-Id: I2566358999234ffe9713c6d0d56f50ef6e777e9b