fix update of java srcdirs after they were already read
It appears that gradle takes a copy of the source directories at
configuration time, because setting them at afterEvaluate had no effect.
This CL changes the plugin to add a closure to the source directories,
this closure is evaluated each time it is required.

This also fixes an issue where the user-specified package translations
were overriding the default translations. This is required because of
https://vanadium-review.googlesource.com/#/c/17487/.

Change-Id: I4b9d0143b8f364c2e048da18dfcbe790dc33bee3