java/naming: Fix java RoutingID toString function.
The java toString function did not convert bytes to a hex string
properly.
Ideally, we should eventually use some Hex package to implement these.

This was breaking vanadium-java-test in the new rpc system because
the new system uses RoutingId comparison as an indication of whether
or not we are talking to a proxy.
It never caused an issue before because non-test java applications
use names when making RPCs, which then relies on go to do the
resolution, so endpoints never have to cross the language boundary.

Change-Id: Iff46af879f8e8e3100b073561a6e258a5afff87f