java: Remove the -builtin_vdlroot flag.
The vdl tool needs access to its standard packages when compiling
and generating code.  It can either find the packages under
VDLROOT/JIRI_ROOT, or it can use the packages built-in to the
tool itself.

The previous behavior was that if -builtin_vdlroot=false (the
default), we'd always look under VDLROOT or JIRI_ROOT.  The new
behavior gets rid of the flag, and changes the behavior as if it
were set to true.  If VDLROOT or JIRI_ROOT are set, we always use
that value, otherwise we extract the built-in vdlroot to a
temporary directory.

The previous behavior was to write out the data as a
base64/gzipped/tarball with line breaks.  The new behavior writes
out a gzipped/tarball as raw binary data, without line breaks.
This is simpler; there's no need for the base64 encoding.

The previous behavior only included *.vdl files in the tarball;
we also need to include vdl.config files, since they affect code
generation for dependant packages.

Added an explicit test to make sure the builtin vdlroot data
contains exactly what we expect; the net result is similar to the
go-generate test (which reminds us when we've forgotten to run
go-generate), but is a bit more localized.

Change-Id: I2bd086e089500e3171aba09cc23289f43c3f748c
MultiPart: 3/6