Baku Toolkit - Blessing seeker fragment
Switching from "trait" + request fragment to a complete blessing seeker
fragment to address lifecycle issues. With the trait + request fragment:
* Trait starts request fragment
* Request fragment sends Account Manager intent
* Account Manager active
* Screen rotates
* Account Manager returns activity result
* Request fragment is destroyed and recreated. The bug is that the trait
  was observing the original request fragment, and has no way of knowing
  the request fragment has changed. The old request fragment actually
  closed its observable without blessings, which triggered a runtime
  error, but anyway without lifecycle management the trait cannot bind
  to the new request fragment.

Lifecycle management could be provided somewhat by fixing the request
fragment to bundle its seeking state and relying on the activity itself
to recreate the BlessedActivityTrait which would then search for the
presence of the request fragment, but that's hackier than just making
the BlessedActivityTrait a fragment with its own lifecycle awareness.

Also went ahead and made the default ErrorReporter a fragment. This
could have been kept as a trait-like object while making the error
reporter for the AccountManagerBlessingsFragment mutable. However, it is
a UI element, and making it a fragment might open the door for resolving
it via the view hierarchy (TODO) instead of dependency injection.

At this time though, the Account Manager itself has some (non-fatal)
bugs around screen rotation. It will repeat menus/screens for each
rotation, but the eventual activity result is fine. I have verified this
is not due to duplicate intents sent from Baku.

Change-Id: I3dafe2f5fb978603cacbda29ebf982c275d1efc9