Java: fix android library builds for bigger projects
In the case of the reader project, when the dependencies grew too large,
gradle would start applying transforms for Sources, Jar files, as well
as native libraries.  We weren't storing our native libraries in the
right location for these transforms and the native library was getting
dropped.  This CL fixes the right location for native libraries.

Also bumped up the vanadium published version to 0.3.

Change-Id: I256d8f9b7b1ccced9a56fd76396985057a4bbabf