Java: run all server methods in a thread pool
Before this change, we were running server methods inside a Go thread
that happens to be running an invoker, which ends up blocking
the Go thread if the server method blocks, which is a no-no.

With this change, we spawn of a new thread (in a thread pool)
for every server request.  We also provide a way for users
to pass in their own scheduling logic if they so choose
(by providing their own Java thread Executor).

MultiPart: 1/2
Change-Id: I74f39adc81c93d77b1bec2fdaba88a1f05ec43d1