Baku Toolkit - Delegating mixins; adding SB ops
Swapping names of mixins and traits to more accurately reflect
nomenclature; traits are now purely interfaces, whereas mixins specify
the behavior. Mixins are now delegated to directly rather than through
accessors, leading to more fluent use at the expense of depending on a
Lombok annotation that the docs say is difficult to support and may be
removed in the future.

Also adding some syntactic sugar for RxTable put/get.

Change-Id: I277c07117c1e0c279ba8b4b15162bf0299b87514