Java: hide the blessing activity behind a simple API
Prior to this CL, the only way to obtain blessings was to:

(1) call our library function to create an Intent,
(2) invoke that intent via startActivityForResult(),
(3) override onActivityResult() to handle the response, and
(4) call our library function to unpack the response.

The reason this was so complicated is that only activities can
receive responses from other activities, and so we couldn't pack
this into a library.

This CL relies on the communication between BlessingActivity
and the invoker taking place through SharedPreferences, thereby
allowing a library to talk to BlessingActivity.  As a result, the
library interface is:

ListenableFuture<Blessings> getBlessings(Context context, String key);

Change-Id: I2db8124a6a4a5eababa720fda750e9af27b12eec