moments, a photo sharing app
Snap and share ephemeral photos with nearby devices.

An example of running, advertising and scanning for multiple services.
User running the app can take photos, building up a list of photos to
share.

Each photo has an "advertise" switch".  The app itself has a "scan"
switch.  When scanning, the app sees and displays photos advertised by
other devices.

Every photo is served by its own service.  This allows for easy creation
of multiple scan targets to exercise discovery code, and allows fine
grained control of photo advertisement.

The involvement of a photo forces use of RPC since a photo is too large
to send as a discovery attribute.  One can turn scanning on/off, and
turn advertising on/off on a per photo basis to see the impact on the UX
of other devices.

Now that we have a few apps in hand, we might want to start a simple
shared lib containing things like V23Manager and SignInActivity.

Change-Id: I31273ff5192da1f86f33fc03ddffdc8a2ab28e99