Baku Toolkit - adding default async behavior
Adds default async behavior for Syncbase asyncs executed in a Baku
context:

* Start every individual Syncbase table command (everything run via
  exec) immediately and report any errors to Baku's onSyncError unless
  an explicit subscription is made.
* Terminate every observable generated from a Syncbase table when the
  Baku context is destroyed.

These defaults should prevent the majority of errors one would see using
these asyncs in Baku. There are cases where the latter restriction is
not necessary and thus limits the power of the APIs, but in theory the
Syncbase client any such asyncs are operating on is itself bound to the
Baku context, and continuing to operate on that Syncbase client or in
response to any ongoing asyncs via that client shouldn't actually make
sense.

Change-Id: Idfd4c97864059f34af98262ac28c94c442821743