Java: change V.init() to support Android lifecycle integration.
After this change, the expected usage pattern on Android is:

VContext mBaseContext;

protected void onCreate(Bundle savedInstanceState) {
    super.onCreate(savedInstanceState);
    mBaseContext = V.init(this);
}

protected void onDestroy() {
    super.onDestroy();
    mBaseContext.cancel();
}

In this new pattern, each activity/service obtains a fresh context upon
calling V.init().  This context will contain all resources (e.g, started
servers, callbacks) that are local to this activity.  Once this context
is canceled, only the resources belonging to that activity are cleared
- other activities' resources are left untouched.

Above behavior allowes the activity to correctly cleanup its state when
destroyed.  Moreover, in the future, it will prevent callbacks from
being invoked on the activity once the context is canceled (i.e., once
onDestroy() is called).  (Not sure if this can be entirely achieved, we
may have to require the users to check if the context is canceled
before running any code.)

MultiPart: 1/2

Change-Id: Ib6a3f3e8f2ab61314868f3a169831a6d140650f0