Java: ensure that all user futures are executed on a desired executor
We want all futures in our API to be executed on the user-specified
executor.  We also want them to fail if the context used to create
them has been canceled.  This is especially important for streaming
methods where we may get a series of successes otherwise even after
the context has been canceled.

The ultimate goal is to ensure that:

1) once a base context is canceled on onDestroy() on Android, there
can be no successful outstanding futures: all outstanding futures
should fail with CancelledException.
2) all user code by default is executed on the UI thread on android.

MultiPart: 1/2
Change-Id: I086139c591e33ee6fe18d96d2d23c7c15772a70f