Baku - Usability improvements
RecyclerView is now a compile dependency rather than provided. Since the
unit of classloading is a class, RecyclerView is required whenever a
collection binding builder is used, even if used to bind to a
ListView. It's thus still not always required, but this requirement
would be surprising to a developer, and it's better to let ProGuard
strip it out later if possible.

Fixed a case where the type requirement on a prefix binding builder
would throw NPE instead of IllegalStateException.

Prefix binding builder and associated adapters now take Comparator
rather than Ordering (which implements Comparator).

Change-Id: I53c6991e265951b1f1634d228fcb8b4e5330992e