android-lib: Library to help with making Vanadium functionality in Android applications remotely inspectable with the "debug browse" command.
This patch introduces a class that starts up a secure network services
and produces principals that can access it to inspect application state
via the "debug browse" command. Hooking this up into the user interface
for each application is still something that the application must deal
with. As an example, I intend to hook this up into the "Moments" demo
application in a subsequent patch.

Change-Id: I87663e865e54f9ac74921f238de62d82784c5248