Cognitive walkthrough 2016.03.07; delombok srcjar
For a sample, BakuActivity with Lombok:
https://paste.googleplex.com/6469180907847680
Delomboked:
https://paste.googleplex.com/4999330490482688

I'm actually not convinced that the Delomboked source adds sufficient
value. I might consider removing the @Generated annotations, but the big
problem is that doing go-to-declaration on, say,
this.mBakuActivityTrait.binder() actually just goes to the interface
spec. (The same can be done by clicking on its entry in the Structure
pane of the Lomboked source, but that's harder to discover.)

Change-Id: I8eaa19298b857a12f878459c16ae8422df018a99