RemoteInspection: Re-use existing server, list out all addresses
Prior to this patch, RemoteInspectors would start up a new network
server. This is fine for apps that only act as Vanadium RPC clients but
if the application contains an RPC server (such as syncbase), then
objects served by that server will not be available via the debug
browser (for example, "Glob" on the server will not list out
the objects served by it).

With this patch, if RemoteInspection is instantiated with a VContext
that has a server associated with it, the same server provided to
the debug browser.

Furthermore, in this change we also list out *all* the addresses
(all names, all endpoints) that the server is available on in the
hope that at least one of them will be accessible from another device.

Change-Id: Ic85e14590f2ba6602c2c32b2bcb3d996644fdf03