DebugFragment: Remove remote inspection.
Remote inspection in applications that run a server (such as in
syncbase) is best initiated with the same VContext that runs the server
so that the services exported by the server can be inspected as well.

For example, for the TODOs application, this was set in:
https://vanadium-review.googlesource.com/#/c/22141/

To avoid confusion between the server started by the application
and a server implicitly created by the DebugFragment, remove the latter.

Change-Id: I80556da158c190c276da1cf90c51fc752f07e492