java: Fix the behavior of addSyncbaseLib when JIRI_ROOT is missing
The fact that VanadiumEnvironment.getVanadiumEnvironment throws an
error when JIRI_ROOT is missing is problematic for the addSyncbaseLib
that is suppose to not fail even in that case. The workaround
implemented in this change is to catch that error and terminate early.
If we end up using only the absolute path for JIRI_ROOT then I'll
simplify the code some more.

Change-Id: Ied3c556b62e6e42d19612cda25603ab1e0c4aa7a