RemoteInspectors: Allow Read/Write methods to be invoked if the caller wants.
With this change, clients of RemoteInspectors can create ones that
allow invitees to invoke Read and Write methods by explicitly passing
in the set of method tags.

The primary motivation for this is some syncbase-on-android
demos so that "debug browse" can list out data (collections, rows,
syncgroups) about syncbase using all the Read methods on the store and
sync interfaces.

Change-Id: I0bff6a4d196464fc579ffdec29c1f85150cb9d38