java: Implement 23 more JNI syncbase internal functions
The functions added in this CL are:

  Database.Destroy
  Database.Exists
  Database.BeginBatch
  Database.Commit
  Database.Abort
  Database.ListCollections
  Database.GetResumeMarker
  Database.ListSyncgroups
  Database.CreateSyncgroup
  Collection.Create
  Collection.Destroy
  Collection.Exists
  Collection.DeleteRange
  Row.Put
  Row.Get
  Row.Exists
  Row.Delete
  Blessings.DebugString
  Blessings.AppBlessingFromContext
  Blessings.UserBlessingFromContext
  Util.Encode
  Util.EncodeId
  Util.NamingJoin

This CL adds idempotent "x.free" methods to types.go and changes the
"x.toFoo" functions to leave the "x" in the same state as "x.free".

This CL also fixes the cExists argument for v23_syncbase_RowExists
to be *C.v23_syncbase_Bool instead of *bool.

MultiPart: 2/2
Change-Id: I5b72bfd8b4754d9fe4ce06f571232dd46da7f383