java: Fix "signature verification failed" error.
When we reinstall a Vanadium Android app, we sometimes get a
"signature verification failed" error. This is because the private
key from/generatedby the KeyStore, doesn't match the signature left
over in the blessingstore in internal storage.

This change makes sure nothing can go wrong, by deleting the private
key from the key store and clearing the internal storage and recreating
a principal, whenever we get a signature verification failed error.

Change-Id: I2cbda6fd66afd6737ac179ec654a26c4de363903