java: Get most of the JNI internal tests to pass again
As expected, the blessing enforcement that was added by v.io/c/22770
and v.io/c/22771 broke the existing JNI tests. This fixes most of them
by using the '...' as a blessing.

The two current failing tests are:

  - UserBlessingFromContext
  - AppBlessingFromContext

Both are failing with a v.io/v23/syncbase/util.FoundNoConventionalBlessings
error.

Change-Id: If48f3382a84a554ef6700a2de9af40445cf5641b