java: Implement the Database.WatchPatterns and Collection.Scan
I wasn't able to see that the onChange is called (which matches what
zinman@ also observed).

This change also contains two fixes to the ScanCallbacks interface:

 - the onValue method is renamed to onKeyValue
 - the onDone is updated to take a VError argument.

Due to some new "inconsistent definitions" I also had to move the
jFindClass from jni_util to jni.go. The bottom of the jni.go and the
jni_lib.go are also sorted.

MultiPart: 2/2
Change-Id: I4fba94a06b0398e46623ec39e1350ff179dec52a