java: Release 2.2.2
Sync with version 2.1.9 (and under) is no longer compatible due to a
recent syncgroup metadata change. Compared to the other revisions in
the 2.1.x line, we cannot refactor client-side code to fix the
problem, so I've decided that we need a minor version bump.

Notably, the TODOs app depends on 2.1.+, so it will be switched to
the 2.2.+ line shortly.

2.2.0: Backwards incompatible metadata change
2.2.1: RPC fix that fixes some "doesn't sync" bugs.
2.2.2: Includes listenForInvites in Java

Change-Id: Iad4f37154972b738ac1d41b3249195c95b630e21