java/syncbase: Login / Presence / User
Modifications to expand upon what it means to login, share one's
presence, and what User data we might be able to return.

The User data will include only a single alias, which depends on
the identity provider. This strikes a good balance between us
using the User's id + identity provider to get enough information
for the blessing, while also giving developer's something usable to
present in the UI.

Change-Id: I1d9db425ebbdf93e184c85643360a74aaf6b07e9