java: Switch the high-level Syncbase API to the new JNI
The DatabaseOptions.rootDir is not yet used and that is causing one
test to fail. I'm going to fix that in a subsequent change because it
will touch also the Go code.

Also note that I temporary disable the test for POJO because it was
trying to call a native function from the v23.so.

Change-Id: If28925bebbafafcf22df407162e8f2626f8a5f64