java: Get the DatabaseOptions.rootDir to work
Note that the rootDir is not exposed in the io.v.syncbase.core. The
purpose of this CL is to get all the high-level API tests from
SyncbaseTest to work. This change also adds the ability to shutdown
syncbase. The high-level API tests are reconfigured to use and empty
list of mounts points speed up the shutdown.

MultiPart: 2/3
Change-Id: Idd810de20c65dc21bac8d8841e64854abbaa4bd3